+++ Sources/Error.swift
@@ -12,11 +12,26 @@ protocol ANSIConvertible : Error, CustomStringConvertible {
 
 extension ANSIConvertible {
   func print() {
-    if isatty(fileno(stderr)) != 0 {
+    // Check if Xcode Colors is installed and enabled.
+    let xcodeColorsEnabled = (getEnvValue("XcodeColors") == "YES")
+    if xcodeColorsEnabled {
+      fputs("\(ansiDescription)\n", stderr)
+    }
+    
+    // Check if we are in any term env and the output is a tty.
+    let termType = getEnvValue("TERM")
+    if let t = termType, t.lowercased() != "dumb" && isatty(fileno(stdout)) != 0 {
       fputs("\(ansiDescription)\n", stderr)
     } else {
       fputs("\(description)\n", stderr)
     }
+  }
+
+  private func getEnvValue(_ key: String) -> String? {
+    guard let value = getenv(key) else {
+      return nil
+    }
+    return String(cString: value)
   }
 }